new guild module and mythic raid leaderboards

+++ app/src/main/java/com/BlizzardArmory/model/warcraft/mythicraid/Entries.kt
@@ -4,10 +4,9 @@ import com.google.gson.annotations.SerializedName
 
 
 data class Entries(
-
     @SerializedName("guild") val guild: Guild,
     @SerializedName("faction") val faction: Faction,
-    @SerializedName("timestamp") val timestamp: Int,
+    @SerializedName("timestamp") val timestamp: Long,
     @SerializedName("region") val region: String,
     @SerializedName("rank") val rank: Int
 )
